style: decouple the deopendencies
