+++ client/src/controller/ILLMController.ts
@@ -1,0 +1,3 @@
+export interface ILLMService {
+    ask(question: string): Promise<string | null>;
+}
