+++ client/src/services/AuthService.ts
@@ -1,8 +1,5 @@
 import { jwtDecode } from 'jwt-decode';
-
-interface DecodedToken {
-  exp: number;
-}
+import { DecodedToken } from '../types/DecodedToken';
 
 export class AuthService {
   static storeToken(credential: string): boolean {
@@ -14,6 +11,18 @@ export class AuthService {
       }
     } catch (err) {
       console.error("Token validation error:", err);
+    }
+    return false;
+  }
+
+  public static signOut() {
+    document.cookie = "authToken=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;";
+  }
+
+  public static handleGoogleResponse(response: unknown): boolean {
+    if (response && typeof response === 'object' && 'credential' in response) {
+      const { credential } = response as { credential: string };
+      return !!(credential && this.storeToken(credential));
     }
     return false;
   }
