--- client/src/services/ILLMService.ts
@@ -1,3 +1,0 @@
-export interface ILLMService {
-  ask(question: string): Promise<string | null>;
-}
