+++ client/src/types/DecodedToken.ts
@@ -1,0 +1,3 @@
+export interface DecodedToken {
+  exp: number;
+}
